tools: Get rid of the embedded package manager
Helm, tilt and flux are currently all installed separately, and we
tell you to do so in the development guide.

Envtest was managed by the package manager - I've replaced it with
setup-envtest, which is a tool that sets this up for you.

There's various other things we need. These were variously installed
in different ways. Now, they're all installed with `@go install
package@version` just before it's needed - as long as you have a go
complier (already required), and you've set up your go bin path to be
part of your PATH (which I've added documentation for), it should mean
that all tools will be installed before they're needed, in the version
that is needed, whereas before you just had to remember to run `rm
.deps; make dependencies` from time to time.

I chose to put the versions in the makefile instead of in the tools
file - it means that the protobuf compiler's dependencies won't
interfere with the dependencies of gitops itself, though it does mean
that we might forget to upgrade them.