tools: Get rid of the embedded package manager
Helm, tilt and flux are currently all installed separately, and we
tell you to do so in the development guide.

We install various things ad-hoc from source in the makefile - this
has the advantage that everybody should already have a go compiler,
and that it's more robust than a bit of ad-hoc bash - for instance, if
we change the dependency version then the dependency will be upgraded
on everybody's system when they check out the code, which wasn't the
case before.

This also puts an FAQ item into the development guide about putting
the go bin path into your PATH.