Add new generic getObject/listObjects endpoints
These can be used to return a generic k8s object with all the
details. It's definitely hacky to put json inside protobuf, but this
leaves us with an envelope where we can set whatever we want, and a
payload that contains data following the relatively stable k8s
schemas.

This isn't used yet.