Use getObjects to render source details
Before this, we had to list every kind of source, filter through all
of them, and try to find the one that matched the request.

This uses the reply from `getObject` to extract the data we care about
instead, by porting the code from `core/server/types`.

This isn't simply duplication - I want to use this in more
places. However, I figured the place that is the most inefficient with
the old code would be the best place to start with this change.