cli: Push client instantiation closer to each command
This makes only the commands that need a client create a client.

This also changes the tests for some of the commands that uses the
client - in practice, this just makes the test test the command
itself, rather than testing that cobra successfully parsed a command
line. I'm sure cobra can parse command lines, so this is better.

This is preliminary work to help do #1992.