gitops run: Don't stop uploading on every single error
Because git doesn't support checking in a directory - only files - it
is a common pattern to have e.g. an empty file called `.keep` in
directories you want to, well, keep.

However, our bucket server gets upset if we try to upload an empty
file and throws an error.

If it ever threw any error, we stopped uploading - we didn't shut
anything down, we just stopped.

This makes it so that a `MissingContentLength` error is quietly
swallowed, and any other error is printed but doesn't stop the process.

This means that instead of uploading ~80 files from the weave-gitops
repository, now it uploads 80,000.